kernel_crashdump: Enforce that the log/dmesg files are not a symlink
This prevents normal users from pre-creating a symlink to the predictable
.crash file, and thus triggering a "fill up disk" DoS attack when the .crash
report tries to include itself.

Also clean up the code to make this easier to read: Drop the "vmcore_root"
alias, move the vmcore and vmcore.log cleanup into the "no kdump" section,
and replace the buggy os.walk() loop with a glob to only catch direct
timestamp subdirectories of /var/crash/.

Thanks to halfdog for discovering this!

CVE-2015-1338
Part of LP #1492570