Fix path traversal vulnerability with hooks execution
Ensure that Package: and SourcePackage: fields loaded from reports do not
contain directories. Until now, an attacker could trick a user into opening a⎵
malicious .crash file containing

  Package: ../../../../some/dir/foo

which would execute /some/dir/foo.py with arbitrary code.

Note: This did not work for SourcePackage: as this would require a⎵
/usr/share/apport/*-hooks/source_/ directory -- but sanity-check the value
anyway as a second line of defence.

Thanks to Donncha O'Cearbhaill for discovering this!

CVE-2016-9950
LP: #1648806