apport-{gtk,kde}: Only offer "Relaunch" for recent /var/crash crashes
It only makes sense to offer relaunching for crashes that just happened and
the apport UI got triggered on those. When opening a .crash file copied from
somewhere else or after the crash happened, this is even actively dangerous as
a malicious crash file can specify any arbitrary command to run.

Note that apport-cli has never offered restarting (as that's not a reasonable
thing to do in a shell), thus does not need to be changed.

Thanks to Donncha O'Cearbhaill for discovering this! 

CVE-2016-9951
LP: #1648806