Remember option so that users can diminish crash interactions
Add and adapt test for a "remember" option, which is reported by the UI,
to tell that we want to store the send/don't send option.
Ensure that Relaunch/Force close/Leave close additional buttons are
changed to a checkbox, and that additional Cancel button is never shown,
as being redundant with don't send if enable reporting is possible.
If enable reporting isn't enabled, leave it to "Continue".
Finally, additional labels to propose Sending the report.
LP: #1778497