Introduce support for non-positional arguments
This makes it much easier to support new kernel features. For
compatibility with old containers, we need to have the global
PID as the 5th argument: The forwarding code parses that argument
and passes the rest of the arguments to the host via a socket.
  
It also checks that there are precisely 6 arguments, though, so we
cannot directly use the new arguments, but need to hack around
a bit to fit them into that space, by introducing --more.
  
(LP: #1732962)